10_linux.in: Merge logic from 10_linux_bls and drop that script
There's a 10_linux_bls snippet that's used in ppc64le machines to generate
a set of menuentry commands by parsing BLS configs in /boot/loader/entries.

This was made because on PowerNV ppc64le machines that boot using the OPAL
firmware, there could be an old version of the Petitboot bootloader that
doesn't have support to parse BLS snippets.

But there is no need to have a separate GRUB script for that and the logic
could just be part of the usual 10_linux snippet.

Also the BLS files could be used directly if the bootloader has support to
parse them, which is the case for GRUB that's used in ppc64le OF or if the
Petitboot used in ppc64le OPAL machines is version 1.8.0 or newer.

So only generate the menuentry commands from the BLS snippets in the case
of OPAL machines with an old Petitboot version.

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>